requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework
Pillow==8.4.0
psycopg==3.0.1
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
flower==0.9.4
celery~=4.4.7
kombu==4.6.11
redis==3.5.3
pytz==2021.3
instagrapi==1.16.2
django-cors-headers==3.6.0
